+++ dotnet/Services/CybersourcePaymentService.cs
@@ -55,12 +55,11 @@ namespace Cybersource.Services
         /// </summary>
         /// <param name="createPaymentRequest"></param>
         /// <returns></returns>
-        public async Task<Payments> BuildPayment(CreatePaymentRequest createPaymentRequest)
+        public async Task<Payments> BuildPayment(CreatePaymentRequest createPaymentRequest, MerchantSettings merchantSettings)
         {
             Payments payment = null;
             try
             {
-                MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
                 string merchantName = createPaymentRequest.MerchantName;
                 string merchantTaxId = string.Empty;
                 bool doCapture = false;
@@ -70,53 +69,10 @@ namespace Cybersource.Services
                     orderSuffix = merchantSettings.OrderSuffix.Trim();
                 }
 
+                (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, createPaymentRequest.MerchantSettings, merchantName);
+
                 string referenceNumber = await _vtexApiService.GetOrderId(createPaymentRequest.Reference, createPaymentRequest.OrderId);
-                if (createPaymentRequest.MerchantSettings != null)
-                {
-                    foreach (MerchantSetting merchantSetting in createPaymentRequest.MerchantSettings)
-                    {
-                        switch (merchantSetting.Name)
-                        {
-                            case CybersourceConstants.ManifestCustomField.CompanyName:
-                                merchantName = merchantSetting.Value;
-                                break;
-                            case CybersourceConstants.ManifestCustomField.CompanyTaxId:
-                                merchantTaxId = merchantSetting.Value;
-                                break;
-                            case CybersourceConstants.ManifestCustomField.CaptureSetting:
-                                if (merchantSetting.Value != null)
-                                {
-                                    doCapture = merchantSetting.Value.Equals(CybersourceConstants.CaptureSetting.ImmediateCapture);
-                                }
-
-                                break;
-
-                            case CybersourceConstants.ManifestCustomField.MerchantId:
-                                if (!string.IsNullOrWhiteSpace(merchantSettings.MerchantId))
-                                {
-                                    merchantSettings.MerchantId = merchantSetting.Value;
-                                }
-
-                                break;
-                            case CybersourceConstants.ManifestCustomField.MerchantKey:
-                                if (!string.IsNullOrWhiteSpace(merchantSettings.MerchantKey))
-                                {
-                                    merchantSettings.MerchantKey = merchantSetting.Value;
-                                }
-
-                                break;
-                            case CybersourceConstants.ManifestCustomField.SharedSecretKey:
-                                if (!string.IsNullOrWhiteSpace(merchantSettings.SharedSecretKey))
-                                {
-                                    merchantSettings.SharedSecretKey = merchantSetting.Value;
-                                }
-
-                                break;
-                        }
-                    }
-                }
-                
-                this.BinLookup(createPaymentRequest.Card.Bin, createPaymentRequest.PaymentMethod, out bool isDebit, out string cardType, out CybersourceConstants.CardType cardBrandName);
+                this.BinLookup(createPaymentRequest.Card.Bin, createPaymentRequest.PaymentMethod, out bool isDebit, out string cardType, out CybersourceConstants.CardType cardBrandName, merchantSettings);
                 payment = new Payments
                 {
                     merchantInformation = new MerchantInformation
@@ -600,12 +556,21 @@ namespace Cybersource.Services
         /// <param name="authenticationTransactionId"></param>
         /// <param name="consumerAuthenticationInformation"></param>
         /// <returns></returns>
-        public async Task<(CreatePaymentResponse, PaymentsResponse)> CreatePayment(CreatePaymentRequest createPaymentRequest, string authenticationTransactionId = null, ConsumerAuthenticationInformation consumerAuthenticationInformation = null)
+        public async Task<(CreatePaymentResponse, PaymentsResponse)> CreatePayment(CreatePaymentRequest createPaymentRequest, string authenticationTransactionId = null, ConsumerAuthenticationInformation consumerAuthenticationInformation = null, MerchantSettings merchantSettings = null)
         {
             CreatePaymentResponse createPaymentResponse = null;
             PaymentsResponse paymentsResponse = null;
             try
             {
+                if (merchantSettings == null)
+                {
+                    merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                    string merchantName = createPaymentRequest.MerchantName;
+                    string merchantTaxId = string.Empty;
+                    bool doCapture = false;
+                    (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, createPaymentRequest.MerchantSettings, merchantName);
+                }
+
                 PaymentData paymentData = await _cybersourceRepository.GetPaymentData(createPaymentRequest.PaymentId);
                 if (paymentData == null)
                 {
@@ -631,13 +596,12 @@ namespace Cybersource.Services
                                 // Need to check if there has already been a successful capture
                                 string referenceNumber = await _vtexApiService.GetOrderId(paymentData.CreatePaymentRequest.OrderId);
                                 string orderSuffix = string.Empty;
-                                MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
                                 if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
                                 {
                                     orderSuffix = merchantSettings.OrderSuffix.Trim();
                                 }
 
-                                SearchResponse searchResponse = await this.SearchTransaction($"{referenceNumber}{orderSuffix}");
+                                SearchResponse searchResponse = await this.SearchTransaction($"{referenceNumber}{orderSuffix}", merchantSettings);
                                 if (searchResponse != null)
                                 {
                                     createPaymentResponse.PaymentId = createPaymentRequest.PaymentId;
@@ -683,7 +647,7 @@ namespace Cybersource.Services
                 }
                 
                 bool isPayerAuth = false;
-                Payments payment = await this.BuildPayment(createPaymentRequest);
+                Payments payment = await this.BuildPayment(createPaymentRequest, merchantSettings);
                 try
                 {
                     ConsumerAuthenticationInformation consumerAuthenticationInformationToCopy = consumerAuthenticationInformation != null ? consumerAuthenticationInformation : paymentData.ConsumerAuthenticationInformation;
@@ -724,7 +688,8 @@ namespace Cybersource.Services
                     });
                 }
 
-                paymentsResponse = await _cybersourceApi.ProcessPayment(payment, createPaymentRequest.SecureProxyUrl, createPaymentRequest.SecureProxyTokensUrl);
+                paymentsResponse = await _cybersourceApi.ProcessPayment(payment, createPaymentRequest.SecureProxyUrl, createPaymentRequest.SecureProxyTokensUrl, merchantSettings);
+
                 if (paymentsResponse != null)
                 {
                     createPaymentResponse = new CreatePaymentResponse();
@@ -748,7 +713,6 @@ namespace Cybersource.Services
                     {
                         try
                         {
-                            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
                             if (!string.IsNullOrWhiteSpace(merchantSettings.CustomNsu))
                             {
                                 createPaymentResponse.Nsu = await this.GetReconciliationId(merchantSettings, createPaymentRequest);
@@ -868,6 +832,10 @@ namespace Cybersource.Services
                 string referenceNumber = await _vtexApiService.GetOrderId(paymentData.CreatePaymentRequest.OrderId);
                 string orderSuffix = string.Empty;
                 MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                string merchantName = paymentData.CreatePaymentRequest.MerchantName;
+                string merchantTaxId = string.Empty;
+                bool doCapture = false;
+                (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, paymentData.CreatePaymentRequest.MerchantSettings, merchantName);
                 if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
                 {
                     orderSuffix = merchantSettings.OrderSuffix.Trim();
@@ -896,7 +864,7 @@ namespace Cybersource.Services
                     }
                 };
 
-                PaymentsResponse paymentsResponse = await _cybersourceApi.ProcessReversal(payment, paymentData.AuthorizationId);
+                PaymentsResponse paymentsResponse = await _cybersourceApi.ProcessReversal(payment, paymentData.AuthorizationId, merchantSettings);
                 if (paymentsResponse != null)
                 {
                     cancelPaymentResponse = new CancelPaymentResponse();
@@ -971,6 +939,19 @@ namespace Cybersource.Services
 
                 string orderSuffix = string.Empty;
                 MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                string merchantName = string.Empty;
+                if (paymentData != null && paymentData.CreatePaymentResponse != null && !string.IsNullOrEmpty(paymentData.CreatePaymentRequest.MerchantName))
+                {
+                    merchantName = paymentData.CreatePaymentRequest.MerchantName;
+                }
+
+                string merchantTaxId = string.Empty;
+                bool doCapture = false;
+                if (paymentData != null && paymentData.CreatePaymentRequest != null && paymentData.CreatePaymentRequest.MerchantSettings != null)
+                {
+                    (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, paymentData.CreatePaymentRequest.MerchantSettings, merchantName);
+                }
+
                 if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
                 {
                     orderSuffix = merchantSettings.OrderSuffix.Trim();
@@ -1117,7 +1098,7 @@ namespace Cybersource.Services
                 }
                 #endregion Custom Payload
 
-                PaymentsResponse paymentsResponse = await _cybersourceApi.ProcessCapture(payment, authId);
+                PaymentsResponse paymentsResponse = await _cybersourceApi.ProcessCapture(payment, authId, merchantSettings);
                 if (paymentsResponse != null)
                 {
                     capturePaymentResponse = new CapturePaymentResponse();
@@ -1135,7 +1116,7 @@ namespace Cybersource.Services
                     else
                     {
                         // Try to get transaction from Cybersource
-                        SearchResponse searchResponse = await this.SearchTransaction($"{referenceNumber}{orderSuffix}");
+                        SearchResponse searchResponse = await this.SearchTransaction($"{referenceNumber}{orderSuffix}", merchantSettings);
                         if (searchResponse != null)
                         {
                             foreach (var transactionSummary in searchResponse.Embedded.TransactionSummaries.Where(transactionSummary => transactionSummary.ApplicationInformation.Applications.Any(ai => ai.Name.Equals(CybersourceConstants.Applications.Capture) && ai.ReasonCode.Equals("100"))))
@@ -1212,6 +1193,10 @@ namespace Cybersource.Services
                 string referenceNumber = await _vtexApiService.GetOrderId(refundPaymentRequest.PaymentId);
                 string orderSuffix = string.Empty;
                 MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                string merchantName = paymentData.CreatePaymentRequest.MerchantName;
+                string merchantTaxId = string.Empty;
+                bool doCapture = false;
+                (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, paymentData.CreatePaymentRequest.MerchantSettings, merchantName);
                 if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
                 {
                     orderSuffix = merchantSettings.OrderSuffix.Trim();
@@ -1247,7 +1232,7 @@ namespace Cybersource.Services
                     };
                 }    
 
-                PaymentsResponse paymentsResponse = await _cybersourceApi.RefundCapture(payment, paymentData.CaptureId);
+                PaymentsResponse paymentsResponse = await _cybersourceApi.RefundCapture(payment, paymentData.CaptureId, merchantSettings);
                 if (paymentsResponse != null)
                 {
                     refundPaymentResponse = new RefundPaymentResponse();
@@ -1367,53 +1352,19 @@ namespace Cybersource.Services
                         payment.orderInformation.lineItems.Add(lineItem);
                     }
 
-                    MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
                     PaymentRequestWrapper requestWrapper = new PaymentRequestWrapper(sendAntifraudDataRequest);
+                    MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
                     string merchantName = string.Empty;
                     string merchantTaxId = string.Empty;
-                    if (sendAntifraudDataRequest.MerchantSettings != null)
-                    {
-                        foreach (MerchantSetting merchantSetting in sendAntifraudDataRequest.MerchantSettings)
-                        {
-                            switch (merchantSetting.Name)
-                            {
-                                case CybersourceConstants.ManifestCustomField.CompanyName:
-                                    merchantName = merchantSetting.Value;
-                                    break;
-                                case CybersourceConstants.ManifestCustomField.CompanyTaxId:
-                                    merchantTaxId = merchantSetting.Value;
-                                    break;
-                                case CybersourceConstants.ManifestCustomField.MerchantId:
-                                    if (!string.IsNullOrWhiteSpace(merchantSettings.MerchantId))
-                                    {
-                                        merchantSettings.MerchantId = merchantSetting.Value;
-                                    }
-
-                                    break;
-                                case CybersourceConstants.ManifestCustomField.MerchantKey:
-                                    if (!string.IsNullOrWhiteSpace(merchantSettings.MerchantKey))
-                                    {
-                                        merchantSettings.MerchantKey = merchantSetting.Value;
-                                    }
-
-                                    break;
-                                case CybersourceConstants.ManifestCustomField.SharedSecretKey:
-                                    if (!string.IsNullOrWhiteSpace(merchantSettings.SharedSecretKey))
-                                    {
-                                        merchantSettings.SharedSecretKey = merchantSetting.Value;
-                                    }
-
-                                    break;
-                            }
-                        }
-                    }
+                    bool doCapture = false;
+                    (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, sendAntifraudDataRequest.MerchantSettings, merchantName);
 
                     requestWrapper.MerchantId = merchantSettings.MerchantId;
                     requestWrapper.CompanyName = merchantName;
                     requestWrapper.CompanyTaxId = merchantTaxId;
                     payment.merchantDefinedInformation = await this.GetMerchantDefinedInformation(merchantSettings, requestWrapper);
 
-                    PaymentsResponse paymentsResponse = await _cybersourceApi.CreateDecisionManager(payment);
+                    PaymentsResponse paymentsResponse = await _cybersourceApi.CreateDecisionManager(payment, merchantSettings);
 
                     sendAntifraudDataResponse = new SendAntifraudDataResponse
                     {
@@ -1470,7 +1421,7 @@ namespace Cybersource.Services
                         sendAntifraudDataResponse.Message = paymentsResponse.ErrorInformation.Message;
                     }
 
-                    if (sendAntifraudDataRequest != null)
+                    if (sendAntifraudDataResponse != null)
                     {
                         await _cybersourceRepository.SaveAntifraudData(sendAntifraudDataRequest.Id, sendAntifraudDataResponse);
                     }
@@ -1523,11 +1474,17 @@ namespace Cybersource.Services
                 return paymentData.CreatePaymentResponse;
             }
 
-            Payments payment = await this.BuildPayment(createPaymentRequest);
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            string merchantName = createPaymentRequest.MerchantName;
+            string merchantTaxId = string.Empty;
+            bool doCapture = false;
+            (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, createPaymentRequest.MerchantSettings, merchantName);
+
+            Payments payment = await this.BuildPayment(createPaymentRequest, merchantSettings);
 
             try
             {
-                paymentsResponse = await _cybersourceApi.SetupPayerAuth(payment, createPaymentRequest.SecureProxyUrl, createPaymentRequest.SecureProxyTokensUrl);
+                paymentsResponse = await _cybersourceApi.SetupPayerAuth(payment, createPaymentRequest.SecureProxyUrl, createPaymentRequest.SecureProxyTokensUrl, merchantSettings);
                 ConsumerAuthenticationInformationWrapper consumerAuthenticationInformationWrapper = new ConsumerAuthenticationInformationWrapper(paymentsResponse.ConsumerAuthenticationInformation);
                 consumerAuthenticationInformationWrapper.CreatePaymentRequestReference = createPaymentRequest.PaymentId;
                 createPaymentResponse = new CreatePaymentResponse
@@ -1572,7 +1529,13 @@ namespace Cybersource.Services
             {
                 try
                 {
-                    Payments payment = await this.BuildPayment(paymentData.CreatePaymentRequest);
+                    MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                    string merchantName = paymentData.CreatePaymentRequest.MerchantName;
+                    string merchantTaxId = string.Empty;
+                    bool doCapture = false;
+                    (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, paymentData.CreatePaymentRequest.MerchantSettings, merchantName);
+
+                    Payments payment = await this.BuildPayment(paymentData.CreatePaymentRequest, merchantSettings);
                     payment.consumerAuthenticationInformation = new ConsumerAuthenticationInformation
                     {
                         ReferenceId = paymentData.PayerAuthReferenceId,
@@ -1580,7 +1543,7 @@ namespace Cybersource.Services
                         ReturnUrl = $"https://{_context.Vtex.Workspace}--{_context.Vtex.Account}.myvtex.com/cybersource/payer-auth-response"
                     };
 
-                    paymentsResponse = await _cybersourceApi.CheckPayerAuthEnrollment(payment, paymentData.CreatePaymentRequest.SecureProxyUrl, paymentData.CreatePaymentRequest.SecureProxyTokensUrl);
+                    paymentsResponse = await _cybersourceApi.CheckPayerAuthEnrollment(payment, paymentData.CreatePaymentRequest.SecureProxyUrl, paymentData.CreatePaymentRequest.SecureProxyTokensUrl, merchantSettings);
                     paymentData.ConsumerAuthenticationInformation = paymentsResponse.ConsumerAuthenticationInformation;
                     await _cybersourceRepository.SavePaymentData(paymentData.PaymentId, paymentData);
 
@@ -1613,13 +1576,19 @@ namespace Cybersource.Services
             
             try
             {
-                Payments payment = await this.BuildPayment(createPaymentRequest);
+                MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                string merchantName = createPaymentRequest.MerchantName;
+                string merchantTaxId = string.Empty;
+                bool doCapture = false;
+                (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, createPaymentRequest.MerchantSettings, merchantName);
+
+                Payments payment = await this.BuildPayment(createPaymentRequest, merchantSettings);
                 payment.consumerAuthenticationInformation = new ConsumerAuthenticationInformation
                 {
                     AuthenticationTransactionId = authenticationTransactionId
                 };
 
-                paymentsResponse = await _cybersourceApi.ValidateAuthenticationResults(payment, createPaymentRequest.SecureProxyUrl, createPaymentRequest.SecureProxyTokensUrl);
+                paymentsResponse = await _cybersourceApi.ValidateAuthenticationResults(payment, createPaymentRequest.SecureProxyUrl, createPaymentRequest.SecureProxyTokensUrl, merchantSettings);
             }
             catch (Exception ex)
             {
@@ -1651,7 +1620,8 @@ namespace Cybersource.Services
 
         public async Task<string> RetrieveAvailableReports(DateTime dtStartTime, DateTime dtEndTime)
         {
-            return await _cybersourceApi.RetrieveAvailableReports(dtStartTime, dtEndTime);
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            return await _cybersourceApi.RetrieveAvailableReports(dtStartTime, dtEndTime, merchantSettings);
         }
 
         public async Task<string> RetrieveAvailableReports(string startTime, string endTime)
@@ -1663,7 +1633,8 @@ namespace Cybersource.Services
 
         public async Task<string> GetPurchaseAndRefundDetails(DateTime dtStartTime, DateTime dtEndTime)
         {
-            return await _cybersourceApi.GetPurchaseAndRefundDetails(dtStartTime, dtEndTime);
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            return await _cybersourceApi.GetPurchaseAndRefundDetails(dtStartTime, dtEndTime, merchantSettings);
         }
 
         public async Task<string> GetPurchaseAndRefundDetails(string startTime, string endTime)
@@ -2579,7 +2550,7 @@ namespace Cybersource.Services
             return merchantDefinedInformationList;
         }
 
-        public void BinLookup(string cardBin, string paymentMethod, out bool isDebit, out string cardType, out CybersourceConstants.CardType cardBrandName)
+        public void BinLookup(string cardBin, string paymentMethod, out bool isDebit, out string cardType, out CybersourceConstants.CardType cardBrandName, MerchantSettings merchantSettings)
         {
             isDebit = false;
             VtexBinLookup vtexBinLookup = _vtexApiService.VtexBinLookup(cardBin).Result;
@@ -2598,7 +2569,7 @@ namespace Cybersource.Services
             }
             else
             {
-                CybersourceBinLookupResponse cybersourceBinLookup = _cybersourceApi.BinLookup(cardBin).Result;
+                CybersourceBinLookupResponse cybersourceBinLookup = _cybersourceApi.BinLookup(cardBin, merchantSettings).Result;
                 if (cybersourceBinLookup != null && cybersourceBinLookup.PaymentAccountInformation != null && cybersourceBinLookup.PaymentAccountInformation.Card != null)
                 {
                     cardType = cybersourceBinLookup.PaymentAccountInformation.Card.Type;
@@ -2678,11 +2649,12 @@ namespace Cybersource.Services
 
         public async Task<RetrieveTransaction> RetrieveTransaction(string requestId)
         {
-            RetrieveTransaction retrieveTransaction = await _cybersourceApi.RetrieveTransaction(requestId);
+            MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+            RetrieveTransaction retrieveTransaction = await _cybersourceApi.RetrieveTransaction(requestId, merchantSettings);
             return retrieveTransaction;
         }
 
-        public async Task<SearchResponse> SearchTransaction(string referenceNumber)
+        public async Task<SearchResponse> SearchTransaction(string referenceNumber, MerchantSettings merchantSettings)
         {
             SearchResponse searchResponse = null;
             CreateSearchRequest searchRequest = new CreateSearchRequest
@@ -2692,7 +2664,7 @@ namespace Cybersource.Services
                 Limit = 2000
             };
 
-            searchResponse = await _cybersourceApi.CreateSearchRequest(searchRequest);
+            searchResponse = await _cybersourceApi.CreateSearchRequest(searchRequest, merchantSettings);
             return searchResponse;
         }
 
@@ -2718,6 +2690,7 @@ namespace Cybersource.Services
                         case "PARTIAL_AUTHORIZED":
                         case "AUTHENTICATION_SUCCESSFUL":
                         case "SOK":
+                        case "ACCEPTED":
                             if (isPayerAuth)
                             {
                                 if (!string.IsNullOrEmpty(paymentsResponse.ConsumerAuthenticationInformation.VeresEnrolled) &&
@@ -2849,12 +2822,16 @@ namespace Cybersource.Services
                             await Task.Delay(6000); // wait for transaction to be available
                             string orderSuffix = string.Empty;
                             MerchantSettings merchantSettings = await _cybersourceRepository.GetMerchantSettings();
+                            string merchantName = createPaymentRequest.MerchantName;
+                            string merchantTaxId = string.Empty;
+                            bool doCapture = false;
+                            (merchantSettings, merchantName, merchantTaxId, doCapture) = await this.ParseGatewaySettings(merchantSettings, createPaymentRequest.MerchantSettings, merchantName);
                             if (!string.IsNullOrEmpty(merchantSettings.OrderSuffix))
                             {
                                 orderSuffix = merchantSettings.OrderSuffix.Trim();
                             }
 
-                            SearchResponse searchResponse = await this.SearchTransaction($"{referenceNumber}{orderSuffix}");
+                            SearchResponse searchResponse = await this.SearchTransaction($"{referenceNumber}{orderSuffix}", merchantSettings);
                             if (searchResponse != null)
                             {
                                 _context.Vtex.Logger.Warn("GetPaymentStatus", null, "Loaded Transactions from Cybersource", new[] { ("searchResponse", JsonConvert.SerializeObject(searchResponse)) });
@@ -3098,6 +3075,71 @@ namespace Cybersource.Services
             {
                 _context.Vtex.Logger.Error("GetItemTaxAmounts", "Tax Total Verification", "Error", ex);
             }
+        }
+
+        /// <summary>
+        /// Parse settings from the Gateway
+        /// </summary>
+        /// <param name="merchantSettings">Merchant Settings from Admin page</param>
+        /// <param name="gatewaySettings">Merchant Settings from Gateway Settings</param>
+        /// <param name="merchantName">Default Merchant Name</param>
+        /// <returns>
+        /// <param name="merchantSettings">Merchant Settings with Gateway override</param>
+        /// <param name="merchantName">merchantName</param>
+        /// <param name="merchantTaxId">merchantTaxId</param>
+        /// <param name="doCapture">doCapture</param>
+        /// </returns>
+        public async Task<(MerchantSettings, string, string, bool)> ParseGatewaySettings(MerchantSettings merchantSettings, List<MerchantSetting> gatewaySettings, string merchantName)
+        {
+            string merchantTaxId = string.Empty;
+            bool doCapture = false;
+
+            if (gatewaySettings != null)
+            {
+                foreach (MerchantSetting merchantSetting in gatewaySettings)
+                {
+                    switch (merchantSetting.Name)
+                    {
+                        case CybersourceConstants.ManifestCustomField.CompanyName:
+                            merchantName = merchantSetting.Value;
+                            break;
+                        case CybersourceConstants.ManifestCustomField.CompanyTaxId:
+                            merchantTaxId = merchantSetting.Value;
+                            break;
+                        case CybersourceConstants.ManifestCustomField.CaptureSetting:
+                            if (merchantSetting.Value != null)
+                            {
+                                doCapture = merchantSetting.Value.Equals(CybersourceConstants.CaptureSetting.ImmediateCapture);
+                            }
+
+                            break;
+
+                        case CybersourceConstants.ManifestCustomField.MerchantId:
+                            if (!string.IsNullOrWhiteSpace(merchantSetting.Value))
+                            {
+                                merchantSettings.MerchantId = merchantSetting.Value;
+                            }
+
+                            break;
+                        case CybersourceConstants.ManifestCustomField.MerchantKey:
+                            if (!string.IsNullOrWhiteSpace(merchantSetting.Value))
+                            {
+                                merchantSettings.MerchantKey = merchantSetting.Value;
+                            }
+
+                            break;
+                        case CybersourceConstants.ManifestCustomField.SharedSecretKey:
+                            if (!string.IsNullOrWhiteSpace(merchantSetting.Value))
+                            {
+                                merchantSettings.SharedSecretKey = merchantSetting.Value;
+                            }
+
+                            break;
+                    }
+                }
+            }
+
+            return (merchantSettings, merchantName, merchantTaxId, doCapture);
         }
 
         private static T DeepCopy<T>(object objToCopy)