+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Fix for account override settings not being used
+
 ## [1.18.0] - 2023-05-22
 
 ### Changed